Final casi 2.0

+++ src/main/java/exparcialg6/demo/Controller/GestorController.java
@@ -3,6 +3,7 @@ package exparcialg6.demo.Controller;
 import exparcialg6.demo.entity.Producto;
 import exparcialg6.demo.entity.Usuario;
 import exparcialg6.demo.repository.ProductoRepository;
+import exparcialg6.demo.repository.ProductoxpedidoRepository;
 import exparcialg6.demo.repository.UsuarioRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpHeaders;
@@ -19,6 +20,7 @@ import org.springframework.web.servlet.mvc.support.RedirectAttributes;
 
 import javax.validation.Valid;
 import java.io.IOException;
+import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
 
@@ -31,9 +33,26 @@ public class GestorController {
     @Autowired
     UsuarioRepository usuarioRepository;
 
+    @Autowired
+    ProductoxpedidoRepository productoxpedidoRepository;
+
     @GetMapping(value = {"", "/"})
-    public String listarProductos(Model model, RedirectAttributes attr) {
+    public String listarProductos(Model model, RedirectAttributes attr, @RequestParam(required = false) Integer pag) {
+        if (pag == null) {
+            pag = 0;
+        }
+        List<Producto> xd = productoRepository.findAll();
+        int a = xd.size() / 7;
+
+
+        ArrayList<Producto> enviar = new ArrayList<>();
+        for (int i = 0; i < 7; i++) {
+            if (xd.size() > i + pag * 7) {
+                enviar.add(xd.get(i + pag * 7));
+            }
+        }
         model.addAttribute("listaProductos", productoRepository.findAll());
+        model.addAttribute("paginacion", a);
         return "producto/listProduct";
     }
 
@@ -93,6 +112,30 @@ public class GestorController {
 
                 return "redirect:/gestor";
             } else {
+                producto.setCodigo(producto.getNombre().substring(0, 2) + "99");
+                if (file.isEmpty()) {
+                    model.addAttribute("msg", "Debe subir un archivo");
+                    return "producto/listProduct";
+
+                }
+                String filename = file.getOriginalFilename();
+                if (filename.contains("..")) {
+                    model.addAttribute("msg", "Debe subir un archivo");
+                    return "producto/listProduct";
+                }
+                try {
+                    producto.setFoto(file.getBytes());
+                    producto.setFotonombre(filename);
+                    producto.setFotocontenttype(file.getContentType());
+                    productoRepository.save(producto);
+                    return "redirect:/gestor";
+
+                } catch (
+                        IOException e) {
+                    e.printStackTrace();
+                    model.addAttribute("msg", "Ocurrio un error");
+                }
+
                 productoRepository.save(producto);
                 attr.addFlashAttribute("msg", "Producto actualizado exitosamente");
                 return "redirect:/gestor";
@@ -100,27 +143,6 @@ public class GestorController {
         }
     }
 
-    @GetMapping("/image/{id}")
-    public ResponseEntity<byte[]> mostrarImagen(@PathVariable("id") int id) {
-        Optional<Producto> opt = productoRepository.findById(id);
-        if (opt.isPresent()) {
-            Producto p = opt.get();
-
-            byte[] imagencomobytes = p.getFoto();
-
-            HttpHeaders httpHeaders = new HttpHeaders();
-            httpHeaders.setContentType(MediaType.parseMediaType(p.getFotocontenttype()));
-
-            return new ResponseEntity<>(
-                    imagencomobytes,
-                    httpHeaders,
-                    HttpStatus.OK);
-
-        } else {
-            return null;
-        }
-
-    }
 
     @PostMapping("/buscarProducto")
     public String buscarProducto(@RequestParam("searchField") String searchField,
@@ -137,12 +159,15 @@ public class GestorController {
                                  RedirectAttributes attr) {
 
         Optional<Producto> producto = productoRepository.findById(id);
-
-        if (producto.isPresent()) {
+        if (producto.isPresent() && productoxpedidoRepository.encontrarPorIdprod(id).isEmpty()) {
             productoRepository.deleteById(id);
             attr.addFlashAttribute("msg", "Producto borrado exitosamente");
+            return "redirect:/gestor";
+        } else {
+            attr.addFlashAttribute("msg", "No se puede borrar un pedido que ya se haya vendido");
+            return "redirect:/gestor";
         }
-        return "redirect:/gestor";
+
 
     }
 }
